# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and Authorization Service API",
        "title": "Auth Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "auth.api.bbapp.in",
    "basePath": "/",
    "paths": {
...
        "/forward-auth": {
            "get": {
                "description": "Validates JWT tokens for Traefik forward auth middleware",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forward Auth"
                ],
                "summary": "Forward authentication for Traefik",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Original request URI",
                        "name": "X-Forwarded-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Original request method",
                        "name": "X-Forwarded-Method",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/forward-auth/customer": {
            "get": {
                "description": "Validates JWT tokens specifically for customer service endpoints with enhanced RBAC logging",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forward Auth"
                ],
                "summary": "Customer service forward authentication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Original request URI",
                        "name": "X-Forwarded-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Original request method",
                        "name": "X-Forwarded-Method",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/forward-auth/product": {
            "get": {
                "description": "Validates JWT tokens specifically for product service endpoints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forward Auth"
                ],
                "summary": "Product service forward authentication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Original request URI",
                        "name": "X-Forwarded-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Original request method",
                        "name": "X-Forwarded-Method",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
...
}